+++ AddOptimization.Contracts/Services/IExternalInvoiceService.cs
@@ -17,9 +17,9 @@ namespace AddOptimization.Contracts.Services
 
         Task<ApiResult<List<ExternalInvoiceResponseDto>>> Search(PageQueryFiterBase filters);
 
-        Task<ApiResult<ExternalInvoiceResponseDto>> FetchInvoiceDetails(int id);
+        Task<ApiResult<ExternalInvoiceResponseDto>> FetchInvoiceDetails(long id);
 
-        Task<ApiResult<ExternalInvoiceResponseDto>> Update(int id, ExternalInvoiceRequestDto model);
+        Task<ApiResult<ExternalInvoiceResponseDto>> Update(long id, ExternalInvoiceRequestDto model);
 
 
     }
