+++ AddOptimization.Data/Entities/ExternalInvoiceDetail.cs
@@ -11,7 +11,7 @@ namespace AddOptimization.Data.Entities
     
     public class ExternalInvoiceDetail : BaseEntityNew<Guid>
     {
-        public int ExternalInvoiceId { get; set; }
+        public long ExternalInvoiceId { get; set; }
         public string Description { get; set; }
         public decimal Quantity { get; set; }
         public decimal UnitPrice { get; set; }
